pierdolić war-machine

+++ libft/ft_strlcat.c
@@ -13,7 +13,7 @@
 
 #include "libft.h"
 
-int	ft_strlcat(char *dst, const char *src, size_t size)
+size_t	ft_strlcat(char *dst, const char *src, size_t size)
 {
 	size_t	dest_l;
 	size_t	src_l;
